% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dt_quantile}
\alias{dt_quantile}
\title{dt_quantile}
\usage{
dt_quantile(x, q, threads = 1L)
}
\arguments{
\item{x}{List}

\item{q}{quantile}

\item{threads}{= 1}
}
\value{
list
}
\description{
dt_quantile
}
