% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FM_stats.R
\name{dt_select}
\alias{dt_select}
\title{dt_select}
\usage{
dt_select(data, select = names(data), add = list(), copy = TRUE)
}
\arguments{
\item{data}{dt}

\item{add}{a named list of column names, values, or formula to specify columns to add}

\item{copy}{TRUE to copy the resulted data.table. If a character vector is supplied, only specified columns will be copied.}

\item{columns}{a character vector of columns to select. If named like c(x = "x1"), the columns in value are renamed to corresponding names}
}
\description{
dt_select
}
