% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{factor_nav}
\alias{factor_nav}
\title{factor_nav, nav = cumsum(factor * ret1)}
\usage{
factor_nav(signals, ret1)
}
\arguments{
\item{signals}{vector}

\item{ret1}{next bar's return}
}
\description{
factor_nav, nav = cumsum(factor * ret1)
}
