% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fast_hist_n}
\alias{fast_hist_n}
\title{fast_hist_n}
\usage{
fast_hist_n(x, bins, min_val = NA_real_, max_val = NA_real_, threads = 1L)
}
\arguments{
\item{x}{vector}

\item{bins}{bin num}

\item{min_val}{min val}

\item{max_val}{max val}

\item{threads}{thread num}
}
\description{
fast_hist_n
}
