% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FM_stats.R
\name{fast_quantile}
\alias{fast_quantile}
\title{fast_quantile}
\usage{
fast_quantile(x, group, qs = c(0.5), threads = 1L)
}
\arguments{
\item{x}{double vector}

\item{group}{int vector}

\item{qs}{double vector contains quantiles}

\item{threads}{default = 1}
}
\description{
fast_quantile
}
