% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fast_quantile_kernel}
\alias{fast_quantile_kernel}
\title{fast_quantile_kernel}
\usage{
fast_quantile_kernel(x, group, qs, threads = 1L)
}
\arguments{
\item{x}{double vector}

\item{group}{int vector}

\item{qs}{double vector contains quantiles}

\item{threads}{default = 1}
}
\description{
fast_quantile_kernel
}
