% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FM.R
\docType{data}
\name{FM}
\alias{FM}
\title{Creates FactorizationMachines model.}
\format{\code{\link{R6Class}} object.}
\usage{
FM
}
\description{
Creates second order Factorization Machines model
}
\section{Usage}{

For usage details see \bold{Methods, Arguments and Examples} sections.
\preformatted{
fm = FM$new(learning_rate = 0.2, rank = 8, lambda_w = 1e-6, lambda_v = 1e-6, task = "classification")
fm$partial_fit(X, y, nthread  = 0, ...)
fm$predict(X, nthread  = 0, ...)
}
}

\section{Methods}{

\describe{
  \item{\code{FM$new(learning_rate = 0.2, rank = 8, lambda_w = 1e-6, lambda_v = 1e-6, task = "classification")}}{Constructor
  for FactorizationMachines model. For description of arguments see \bold{Arguments} section.}
  \item{\code{$partial_fit(X, y, nthread  = 0, ...)}}{fits/updates model given input matrix \code{X} and target vector \code{y}.
  \code{X} shape = (n_samples, n_features)}
  \item{\code{$predict(X, nthread  = 0, ...)}}{predicts output \code{X}}
  \item{\code{$coef()}}{ return coefficients of the regression model}
  \item{\code{$dump()}}{create dump of the model (actually \code{list}) with current model parameters}
}
}

\section{Arguments}{

\describe{
 \item{fm}{\code{FTRL} object}
 \item{X}{Input sparse matrix - native format is \code{Matrix::RsparseMatrix}.
 If \code{X} is in different format, model will try to convert it to \code{RsparseMatrix}
 with \code{as(X, "RsparseMatrix")} call}
 \item{learning_rate}{learning rate for AdaGrad SGD}
 \item{rank}{rank of the latent dimension in factorization}
 \item{lambda_w}{regularization parameter for linear terms}
 \item{lambda_v}{regularization parameter for interactions terms}
 \item{n_features}{number of features in model (number of columns in expected model matrix) }
 \item{task}{ \code{"regression"} or \code{"classification"}}
}
}

\keyword{datasets}
