% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FM_stats.R
\name{fm_quantile}
\alias{fm_quantile}
\title{fm_quantile
plot using ggplot(q, aes(x = qv)) + geom_line(aes(y = density_x))}
\usage{
fm_quantile(
  x,
  q = c(0, 0.005, 0.01, 0.05, (1L:9L) * 0.1, 0.95, 0.99, 0.995, 1)
)
}
\arguments{
\item{x}{numeric vector}

\item{q}{quantile vector}
}
\description{
fm_quantile
plot using ggplot(q, aes(x = qv)) + geom_line(aes(y = density_x))
}
