% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fm_zero_ratio}
\alias{fm_zero_ratio}
\title{fm_zero_ratio}
\usage{
fm_zero_ratio(x, precision = 1e-09)
}
\arguments{
\item{x}{vector}

\item{precision}{precision}
}
\value{
zero_ratio
}
\description{
fm_zero_ratio
}
