% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GBM_ohlc}
\alias{GBM_ohlc}
\title{general brownian motion}
\usage{
GBM_ohlc(ukey = 1L, start_price = 100, n = 1000L, DataDate = 1L)
}
\arguments{
\item{ukey}{ukey}

\item{start_price}{start_price}

\item{n}{length}

\item{DataDate}{date}
}
\value{
list
}
\description{
general brownian motion
}
