% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_binned_stats}
\alias{get_binned_stats}
\title{get_binned_stats}
\usage{
get_binned_stats(x, y, n_bin, threads = 1L)
}
\arguments{
\item{x}{vector, bin divided by x, not y}

\item{y}{vector}

\item{n_bin}{bin num}

\item{threads}{thread num}
}
\description{
get_binned_stats
}
