% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FM_stats.R
\name{list_get}
\alias{list_get}
\title{list_get}
\usage{
list_get(data, x, default = stop("object '", x, "' not found"))
}
\arguments{
\item{data}{a list or data.frame}

\item{x}{name or formula}

\item{default}{value to return if x is not present in data}
}
\description{
list_get
}
