% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FM_invest.R
\name{longest_drawdown}
\alias{longest_drawdown}
\title{Longest drawdown measures}
\usage{
longest_drawdown(x)
}
\arguments{
\item{x}{numeric vector, cumulative PnL with unit capital}
}
\description{
Longest drawdown measures
}
