% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FM_invest.R
\name{max_drawdown}
\alias{max_drawdown}
\title{Maximal drawdown measures}
\usage{
max_drawdown(x)
}
\arguments{
\item{x}{numeric vector, cumulative PnL with unit capital}
}
\description{
Maximal drawdown measures
}
