% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pcor}
\alias{pcor}
\title{pcor}
\usage{
pcor(x, y = NULL, x_sign = 0L, y_sign = 0L, threads = 1L)
}
\arguments{
\item{x}{List}

\item{y}{numeric vector}

\item{x_sign}{filter x}

\item{y_sign}{filter y}
}
\value{
list
}
\description{
pcor
}
