% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FM_fit.R
\name{R2}
\alias{R2}
\title{calc R2 between}
\usage{
R2(pred, obs, formula = "corr", na.rm = TRUE)
}
\arguments{
\item{pred}{numeric vector}

\item{obs}{numeric vector}

\item{formula}{formula}

\item{na.rm}{default TRUE}
}
\description{
calc R2 between
}
