% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rank}
\alias{rank}
\title{rank in place. will filter NAN, normalize to [0, 1], NAN will remain NAN}
\usage{
rank(x)
}
\arguments{
\item{x}{vector}
}
\description{
rank in place. will filter NAN, normalize to [0, 1], NAN will remain NAN
}
