% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FM_stats.R
\name{rank_score}
\alias{rank_score}
\title{rank_score}
\usage{
rank_score(x, ties.method = "average", ...)
}
\arguments{
\item{x}{vec}

\item{ties.method}{see data.table::frank(ties.method=)}

\item{...}{additional parameters passed to data.table::frank}
}
\description{
Rank score maps the minimum value of an input vector to -1 and maximum value to 1 according to their rankings
}
