% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcor}
\alias{rcor}
\title{rcor}
\usage{
rcor(x, y = NULL, x_sign = 0L, y_sign = 0L)
}
\arguments{
\item{x}{List}

\item{y}{numeric vector}

\item{x_sign}{filter x}

\item{y_sign}{filter y}
}
\value{
list
}
\description{
rcor
}
