% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FM_stats.R
\name{stats.default}
\alias{stats.default}
\title{stats default method}
\usage{
\method{stats}{default}(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Arguments passed to or from other methods}
}
\description{
stats default method
}
