% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{t0_nav}
\alias{t0_nav}
\title{t0_nav}
\usage{
t0_nav(signals, ret1, long_t, short_t, cost = 0)
}
\arguments{
\item{signals}{vector}

\item{ret1}{next bar's return}

\item{long_t}{>= threshold, open long}

\item{short_t}{<= threshold, open short}

\item{cost}{cost}
}
\description{
t0_nav
}
