% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vec_scale}
\alias{vec_scale}
\title{scale vector based on given mean/sd}
\usage{
vec_scale(x, mean, sd)
}
\arguments{
\item{x}{vector}

\item{mean}{target mean}

\item{sd}{target sd}
}
\description{
scale vector based on given mean/sd
}
