% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vec_scale_xy}
\alias{vec_scale_xy}
\title{scale vector based on given y's mean/sd}
\usage{
vec_scale_xy(x, y)
}
\arguments{
\item{x}{vector}

\item{mean}{target mean}

\item{sd}{target sd}
}
\description{
scale vector based on given y's mean/sd
}
