% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FM_stats.R
\name{dt_fill}
\alias{dt_fill}
\title{dt_fill}
\usage{
dt_fill(data, columns, selector, fill)
}
\arguments{
\item{data}{dt}

\item{columns}{columns}

\item{selector}{selector}

\item{fill}{fill}
}
\description{
dt_fill
}
